script/requirements.txt
pandas>=2.0
openpyxl>=3.1
psycopg2-binary>=2.9
akshare>=1.13
